♻️ 重构节点移动控制

+++ src/core/controller/concrete/ControllerCutting.tsx
@@ -59,6 +59,11 @@ ControllerCutting.mouseup = (event: MouseEvent) => {
   if (event.button !== 2) {
     return;
   }
+  if (!Stage.isCutting) {
+    return;
+  }
+  Stage.isCutting = false;
+  
   NodeManager.deleteNodes(Stage.warningNodes);
   Stage.warningNodes = [];
